feat!: add nested objects support
Added naive support for nested objects for parsing. The schema syntax
has changed and now if a property is an object, then it is considered a
nested object.

Previously - this meant a more detailed setting of the property.

BREAKING CHANGE: now more fine-tuning is not available through a normal
object. For this, you need to use schemaProperty().