feat: update behavior of optional props in nested objects
Updated behavior for objects where all keys are optional. Now the
parent of these keys will not be deleted. And if it doesn't exist, an
empty object will be created in its place to support the structure.

Previously, if an object contained optional properties that didn't, it
was removed. This led to a break in the hierarchy of nested objects.